pglookout: Initial version of pglookout, a replication monitoring handler
The purpose of pglookout is to monitor PG database nodes, and their
replication status, and act according to that status.

This can mean for example calling a set failover command to
promote a new master in case the previous one goes missing.

pglookout supports two different nodetypes, ones that are installed
on the db nodes themselves, and observer nodes that can be installed
anywhere. The purpose of the pglookout nodes on the DB nodes is to
monitor the replication status of the cluster, and act accordingly,
the observers have a more limited remit, they just observe the cluster
status.