pglookout: Add support for starting to follow new master PG nodes
This commit adds new config keys: autofollow, pg_start_command and
pg_stop command and pg_data_directory. When these are configured
correctly, you can have pglookout force PG to start following
a replication cluster's new master.

This is useful in a scenario where you have a master and two standbys,
the master node dies and one of the standby is promoted in it's stead.
The standby that wasn't furthest on, will automatically start to follow
the other standby that got promoted.