autofollow: use primary_conninfo_template when updating recovery.conf
Previously autofollow would try to replace the previous master's address in
recovery.conf with a new master address whenever it saw the cluster state
change.  This was problematic as it did not account for recovery.conf
without a primary_conninfo (this can be the case after restoring a system
from backups without a live master server, using for example pghoard) and
also did not check whether recovery.conf actually matched the current
cluster state so it was possible for pglookout to not trigger autofollow if
a failover had happened before pglookout was started.

Fix these cases by requiring a new `primary_conninfo_template` configuration
option whenever `autofollow` is used.  The template is used to generate a
new primary_conninfo setting for recovery.conf when master changes; the
template is also used to check that the line in recovery.conf matches
current cluster master at pglookout startup.