failover: perform failover if master node disappears or never appears
If we don't see a master node in the cluster before failover_timeout seconds
have passed since the list of cluster nodes was last refreshed perform a
failover to promote one of the standby nodes.  This is useful when a cluster
has been brought up from backups and there is no existing master node.

If a master node has been present in the cluster and later on disappears
from both connections and configuration we should perform an immediate
failover decision as there is no chance of reconnecting to a node that's no
longer in our configuration.  This speeds up promotion when a node is
deleted from the cluster.