tests: wipe the temporary directory at teardown
Each test case creates a temporary directory for storing a statefile +
optionally some other things; previously we only ever deleted the statefile
but would leave the directory behind.  Let's wipe the entire directory tree
at teardown and prefix it with something that makes it easy to recognize if
it gets left behind because of a failure.

In Python 3 we could use tempfile.TemporaryDirectory, but unfortunately
that's not available on Python 2 which we still support.