timestamps: fix second calculation between datetime objects
timedelta.seconds only returns the seconds-part of the delta and doesn't
consider days (or microseconds).  Use the total_seconds() function instead
which takes all parts of the delta in account.  This fixes cases where we
had > 1 day replication lag and could have incorrectly returned low results.
Usually this was probably not an issue as replication lag rarely grows
beyond a day, and even if it does the window where previous day's time
matches current time is short.

NOTE: we can't use total_seconds() directly from timedelta objects as it's
not available on Python 2.6.  Implement a compat function in common.py.