remote_conns: the key is just a name for the connection, not a host to use
Don't require the key of `remote_conns` entries to have a meaning, instead
parse the host and port specifications of every remote_conn entry when they
are needed for something, which is currently just setting a new
primary_conninfo entry in recovery.conf.

Because they don't have to be hostnames anymore rename the variables
internally from `host` or `hostname` to `instance`.