failover: select the furthest replicated node with the highest identifier
If there are multiple nodes with the same replication positions pick the one
with the "highest" name to make sure pglookouts running on all standbys make
the same decision.  The rationale for picking the "highest" node is that
there's no obvious way for pglookout to decide which of the nodes is "best"
beyond looking at replication positions, but picking the highest id supports
environments where nodes are assigned identifiers from an incrementing
sequence identifiers and where we want to promote the latest and greatest
node.  In static environments node identifiers can be priority numbers, with
the highest number being the one that should be preferred.