pglookout: check cluster state concurrently from all endpoints
Previously we did this sequentially with 5 second timeouts, meaning
if you had a large cluster where a large number of PostgreSQL
cluster were unavailable, it might take a long amount of time
to get up to date state information.

Now since this is done in parallel we should get up to date cluster
state within 5 seconds or so regardless of cluster size or aliveness.