Ensure master WAL position is returned correctly
For the master node we want to track the current position instead of the
most recent replication position. If fetching replication position
succeeded but the second query for getting current position failed, the
master's last replication position was returned instead of it's current
position. This could make standbys to possibly seem to be ahead of the
master even when that was not the case.