Issue 53: stopping of primary when it becomes isolated
Create two new config keys:
  * minimum_replicating_nodes (default: disabled)
    The minimum number of nodes the master should be replicating with.
    If the number of replicating nodes falls below this mark, then
    too_few_replicating_nodes_command is called if defined.
  * too_few_replicating_nodes_command (default: disabled)
    When the number of replicating nodes on the master falls
    below minimum_replicating_nodes, this command is called.
    Note that it will be called repeatedly as long as this condition continues.

The query that is run on the master node is enhanced to count the number
of records in pg_stat_replication; this is stored in in the field
pg_master_replication_connections.

The too_few_replicating_nodes_command can simply stop the master, or do
more checking first, or whatever. Because the script is called repeatedly,
if you want to stop the master, you should first test whether it is running,
to avoid pg_ctl giving errors.