new key to the config: poll_observers_on_warning_only
this key allows the polling of observers only when pglookout notices
that it is in warning mode.
This is useful when the observer is a single entity observing many
different clusters. And in that case pglookout needs to make sure
the db_poll has happened at-least once to retrieve data from observers.

In pglookout's producer-consumer architecture, consumer thread makes
the decision based on polled data. The consumer relies on timeout along
with ping from producer. Because of timeouts the consumer might
accidentally rely on stale data from the producer. And one way to
indicate staleness is when consumer is in warning mode and even though
observers are configured their polling information is not
available/updated.

Once the option is set to True, observers will be polled and in warning
state only. Therefore the consumer needs to make sure it is consuming
data newer than utcnow)(trakced in in observer_state_newer_than). And
waits for latest observer data relying on the fetch_time of the
observer. However this wait period is limited to db_poll_interval.

This in combination allows consumer to have atleast one
round before moving from warning->critical and not act like a trigger
(direct from normal to critical). This trigger is seen when the node is
back from a being heavily loaded and replication was stalled.

On a subsequent successful poll-monitor loop if replication begins
normally consumer should be able to detect it.