Do not consider master to be down if txid_current fails
In some cases executing txid_current() fails or at least does not
complete within the 5 second timeout. Previously that caused the master
to be considered unresponsive, which is quite incorrect. Fetch the
required master WAL LSN position via separate query first so that the
txid_current() problem can be ignored as we have all the information
that we must get from master.