Reload config on main thread
The config reload implementation is rather complex and doing that from
signal handler is unsafe. Especially the part of using the queue is
problematic as the non-re-entrant lock of the queue may already be
held. Move the config reload to happen as part of the main loop to
avoid any locking issues.